individual data loader
Former-commit-id: fc6737eebeea88ff887b7d2fd2cef15a89d4cc1c

+++ texar/modules/decoders/transformer_decoders.py
@@ -113,7 +113,7 @@ class TransformerDecoder(ModuleBase):
                     dec = layers.normalize(dec)
         self.dec = dec
         self.logits = tf.layers.dense(dec, self._vocab_size)
-        self.preds = tf.to_int64(tf.argmax(self.logits, axis=-1))
+        self.preds = tf.to_int32(tf.argmax(self.logits, axis=-1))
 
         if not self._built:
             self._add_internal_trainable_variables()
